+++ reVX/least_cost_xmission/friction_barrier_builder.py
@@ -9,11 +9,11 @@ from typing_extensions import TypedDict, Required
 
 import numpy as np
 import numpy.typing as npt
-from reVX.least_cost_xmission.config.constants import DEFAULT_DTYPE
 
-from reVX.least_cost_xmission.masks import MaskArr, Masks
-from reVX.least_cost_xmission.trans_layer_io_handler import TransLayerIoHandler
 from reVX.least_cost_xmission.utils import rasterize
+from reVX.least_cost_xmission.masks import MaskArr, Masks
+from reVX.least_cost_xmission.config.constants import DEFAULT_DTYPE
+from reVX.least_cost_xmission.transmission_layer_io_handler import TransLayerIoHandler
 
 logger = logging.getLogger(__name__)
