+++ reVX/least_cost_xmission/json_config.py
@@ -7,7 +7,7 @@ from typing import Optional, Dict, List, Union
 from pydantic import BaseModel, DirectoryPath, FilePath, NewPath
 
 from reVX.least_cost_xmission.friction_barrier_builder import FBLayerConfig
-from reVX.least_cost_xmission.offshore_cost_creator import BinConfig
+from reVX.least_cost_xmission.wet_cost_creator import BinConfig
 
 # Keys are GeoTIFF file paths to use with the FBLayerConfig to use for that
 # GeoTIFF
