+++ reVX/least_cost_xmission/utils.py
@@ -4,16 +4,20 @@ Least cost transmission utilities
 import logging
 from typing import Optional, Union, Dict
 
-import numpy.typing as npt
+import pandas as pd
 import geopandas as gpd
-from rasterio import features
-from rasterio.warp import reproject
+import numpy.typing as npt
+from shapely.geometry import Point, LineString
 
-from .trans_layer_io_handler import Profile
+import rasterio as rio
+from rasterio import features
+
+from .transmission_layer_io_handler import Profile
 
 VECTOR_CACHE: Dict[str, gpd.GeoDataFrame] = {}
 
 logger = logging.getLogger(__name__)
+
 def rasterize(fname: str, profile: Profile,
               buffer_dist: Optional[float] = None,
               all_touched: bool = False,
@@ -81,3 +85,66 @@ def rasterize(fname: str, profile: Profile,
 
     logger.debug('Rasterizing complete')
     return rasterized
+
+
+def convert_pois_to_lines(poi_csv_f: str, template_f: str, out_f: str):
+    """
+    Convert POIs in CSV to lines and save in a geopackage as substations. Also
+    create a fake transmission line to connect to the substations.
+
+    Parameters
+    ----------
+    poi_csv_f
+        Path to CSV file with POIs in it
+    template_f
+        Path to template raster with CRS to use for geopackage
+    out_f
+        Path and file name for geopackage
+    """
+    logger.debug('Converting POIs in {} to lines in {}'
+                .format(poi_csv_f, out_f))
+    with rio.open(template_f) as ras:
+        crs = ras.crs
+
+    df = pd.read_csv(poi_csv_f)[['POI Name', 'State', 'Voltage (kV)', 'Lat',
+                                 'Long']]
+
+    pts = gpd.GeoDataFrame(df, geometry=gpd.points_from_xy(df.Long, df.Lat))
+    pts = pts.set_crs('EPSG:4326')
+    pts = pts.to_crs(crs)
+
+    # Convert points to short lines
+    new_geom = []
+    for pt in pts.geometry:
+        end = Point(pt.x + 50, pt.y + 50)
+        line = LineString([pt, end])
+        new_geom.append(line)
+    lines = pts.set_geometry(new_geom, crs=crs)
+
+    # Append some fake values to make the LCP code happy
+    lines['ac_cap'] = 9999999
+    lines['category'] = 'Substation'
+    lines['voltage'] = 500  # kV
+    lines['trans_gids'] = '[9999]'
+
+    # add a fake trans line for the subs to connect to to make LCP code happy
+    trans_line = pd.DataFrame(
+        {
+            'POI Name': 'fake',
+            'ac_cap': 9999999,
+            'category': 'TransLine',
+            'voltage': 500,  # kV
+            'trans_gids': None
+        },
+        index=[9999]
+    )
+
+    trans_line = gpd.GeoDataFrame(trans_line)
+    geo = LineString([Point(0, 0), Point(100000, 100000)])
+    trans_line = trans_line.set_geometry([geo], crs=crs)  # type: ignore
+
+    pois: gpd.GeoDataFrame = pd.concat([lines, trans_line])
+    pois['gid'] = pois.index
+
+    pois.to_file(out_f, driver="GPKG")
+    logger.debug('Complete')
