+++ reVX/least_cost_xmission/wet_cost_creator.py
@@ -1,5 +1,5 @@
 """
-Create offshore costs and save to GeoTIFF.
+Create wet (offshore) costs and save to GeoTIFF.
 """
 import logging
 from typing import List
@@ -9,7 +9,7 @@ import numpy as np
 import numpy.typing as npt
 from reVX.least_cost_xmission.config.constants import DEFAULT_DTYPE, WET_COSTS_TIFF
 
-from reVX.least_cost_xmission.trans_layer_io_handler import TransLayerIoHandler
+from reVX.least_cost_xmission.transmission_layer_io_handler import TransLayerIoHandler
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +24,7 @@ class BinConfig(TypedDict, total=False):
     cost: Required[float]
 
 
-class OffshoreCostCreator:
+class WetCostCreator:
     """
     Create offshore costs and save to GeoTIFF.
     """
@@ -37,8 +37,8 @@ class OffshoreCostCreator:
         """
         self._io_handler = io_handler
 
-    def build_offshore_costs(self, bathy_tiff: str, bins: List[BinConfig],
-                            out_filename = WET_COSTS_TIFF):
+    def build_wet_costs(self, bathy_tiff: str, bins: List[BinConfig],
+                        out_filename = WET_COSTS_TIFF):
         """
         Build complete offshore costs. This is currently very simple. In the
         future, costs will also vary with distance to port.
